@@ -23,7 +23,7 @@
 		"lz-string": "^1.5.0",
 		"prettier": "^3.3.2",
 		"prettier-plugin-svelte": "^3.2.5",
-		"svelte": "^4.2.18",
+		"svelte": "5.0.0-next.166",
 		"svelte-check": "^3.8.3",
 		"tslib": "^2.6.3",
 		"typescript": "^5.5.2",
